feat(#43): extract Xml interface
