+++ src/main/java/com/github/lombrozo/xnav/Xml.java
@@ -21,104 +21,20 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+
 package com.github.lombrozo.xnav;
 
-import java.io.StringWriter;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Stream;
-import javax.xml.transform.OutputKeys;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerConfigurationException;
-import javax.xml.transform.TransformerException;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
-import org.w3c.dom.Document;
 import org.w3c.dom.Node;
-import org.w3c.dom.NodeList;
 
 /**
  * XML abstraction over an XML document.
- * This class is thread-safe.
+ * All implementations of this interface are thread-safe.
  *
  * @since 0.1
  */
-final class Xml {
-
-    /**
-     * Transformer factory.
-     */
-    private static final TransformerFactory TFACTORY = TransformerFactory.newInstance();
-
-    /**
-     * Actual XML document node.
-     */
-    private final Node inner;
-
-    /**
-     * Ctor.
-     *
-     * @param xml XML document as a string.
-     */
-    Xml(final String xml) {
-        this(new StringNode(xml).toNode());
-    }
-
-    /**
-     * Ctor.
-     *
-     * @param node XML document node.
-     */
-    Xml(final Node node) {
-        this.inner = node;
-    }
-
-    @Override
-    public String toString() {
-        try {
-            final Transformer transformer = Xml.TFACTORY.newTransformer();
-            transformer.setOutputProperty(OutputKeys.VERSION, "1.0");
-            transformer.setOutputProperty(OutputKeys.ENCODING, "UTF-8");
-            if (!(this.inner instanceof Document)) {
-                transformer.setOutputProperty(OutputKeys.OMIT_XML_DECLARATION, "yes");
-            }
-            final StringWriter writer = new StringWriter();
-            transformer.transform(new DOMSource(this.inner), new StreamResult(writer));
-            return writer.toString();
-        } catch (final TransformerConfigurationException econf) {
-            throw new IllegalStateException(
-                String.format(
-                    "Failed to configure Transformer for printing XML to a string: %s", this
-                ),
-                econf
-            );
-        } catch (final TransformerException exception) {
-            throw new IllegalStateException(
-                String.format("Failed to transform XML: %s to a human-readable string", this),
-                exception
-            );
-        }
-    }
-
-    @Override
-    public boolean equals(final Object obj) {
-        final boolean result;
-        if (this == obj) {
-            result = true;
-        } else if (obj == null || getClass() != obj.getClass()) {
-            result = false;
-        } else {
-            final Xml other = (Xml) obj;
-            result = this.inner.isEqualNode(other.inner);
-        }
-        return result;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(this.inner);
-    }
+public interface Xml {
 
     /**
      * Get a child node by its name.
@@ -126,22 +42,7 @@ final class Xml {
      * @param element Element name.
      * @return Child.
      */
-    Xml child(final String element) {
-        synchronized (this.inner) {
-            final NodeList nodes = this.inner.getChildNodes();
-            final int length = nodes.getLength();
-            for (int idx = 0; idx < length; ++idx) {
-                final Node child = nodes.item(idx);
-                if (child.getNodeType() == Node.ELEMENT_NODE
-                    && child.getNodeName().equals(element)) {
-                    return new Xml(child);
-                }
-            }
-            throw new IllegalStateException(
-                String.format("Element '%s' not found in '%s'", element, this)
-            );
-        }
-    }
+    Xml child(String element);
 
     /**
      * Get an attribute by its name.
@@ -149,79 +50,40 @@ final class Xml {
      * @param name Attribute name.
      * @return Attribute.
      */
-    Optional<Xml> attribute(final String name) {
-        synchronized (this.inner) {
-            final Node item = this.inner.getAttributes().getNamedItem(name);
-            final Optional<Xml> result;
-            if (Objects.nonNull(item)) {
-                result = Optional.of(new Xml(item));
-            } else {
-                result = Optional.empty();
-            }
-            return result;
-        }
-    }
+    Optional<Xml> attribute(String name);
 
     /**
      * Get the text of the current node.
      *
      * @return Text of the node.
      */
-    Optional<String> text() {
-        synchronized (this.inner) {
-            final Optional<String> result;
-            if (this.inner.getNodeType() == Node.DOCUMENT_NODE) {
-                result = Optional.of("");
-            } else if (this.inner.getNodeType() == Node.ATTRIBUTE_NODE) {
-                result = Optional.of(this.inner.getNodeValue());
-            } else {
-                result = Optional.of(this.inner).map(Node::getTextContent);
-            }
-            return result;
-        }
-    }
+    Optional<String> text();
 
     /**
      * Get children of the current node.
      *
      * @return Children.
      */
-    Stream<Xml> children() {
-        synchronized (this.inner) {
-            final NodeList nodes = this.inner.getChildNodes();
-            final int length = nodes.getLength();
-            return Stream.iterate(0, idx -> idx + 1)
-                .limit(length)
-                .map(nodes::item)
-                .filter(Objects::nonNull)
-                .map(Xml::new);
-        }
-    }
+    Stream<Xml> children();
 
     /**
      * Get the name of the node.
      *
      * @return Node name.
      */
-    String name() {
-        return this.inner.getNodeName();
-    }
+    String name();
 
     /**
      * Copy the XML document.
      *
      * @return Copy of the document.
      */
-    Xml copy() {
-        return new Xml(this.inner.cloneNode(true));
-    }
+    Xml copy();
 
     /**
      * Get the actual node.
      *
      * @return Node.
      */
-    Node node() {
-        return this.inner;
-    }
+    Node node();
 }
